+++ test_server.py
@@ -159,6 +159,7 @@ def main() -> None:
         skip_signature_validation=False,
         hmac_fetcher=lookup_secret_key,
         max_presign_url_usage_attempts=1,  # Set the maximum URL attempts for presigned URLs
+        metrics_port=9090,
     )
 
     start_server(ra)